Hoist inventory call up to AutomationDetail parent
